Split: take away events from __init__.py
